Unregister unused moving space before starting concurrent compaction
For the next GC cycle's mremap to succeed it is important to register
the entire moving space with userfaultfd. The downside is that we get
userfaults for the new TLABs.
We can eliminate these by mapping a zero-page in the unused portion of
the moving space and then unregistering the unused portion. This takes
care of the kernel's constraint that otherwise makes mremap fail the
next time.

Test: manual
Bug: 160737021
Change-Id: Ibae6174e9af82537ff6b6b13eb6b6dc53d5b8cb9
Signed-off-by: Dmitrii <bankersenator@gmail.com>